Add tests for all reduce methods
